:arrow_up: added enpoint getUserByID, also functions and queryBuilders were separated from index
